+++ tests/testthat/test-folders.R
@@ -198,6 +198,47 @@ with_mock_crunch({
             "]}"
         )
     })
+
+    test_that("rename a folder", {
+        # via setName
+        expect_PATCH(
+            ds %>% cd("Group 2") %>% setName("Group 2 New Name"),
+            "https://app.crunch.io/api/datasets/1/folders/2/",
+            '{"name":"Group 2 New Name"}'
+        )
+
+        # but also trying to move only sends name changes
+        expect_POST(
+            ds %>% cd("/") %>% mv("Group 2", "Group 2 New Name"),
+            "https://app.crunch.io/api/datasets/1/folders/",
+            '{"element":"shoji:catalog","body":{"name":"Group 2 New Name"}}'
+        )
+    })
+
+    test_that("rename objects inside a folder", {
+        expect_PATCH(
+            ds %>%
+                cd("Group 1") %>%
+                setNames(c("Year of birth", "Nested folder", "FTW! textvar")),
+            "https://app.crunch.io/api/datasets/1/folders/1/",
+            '{"element":"shoji:catalog","index":{"https://app.crunch.io/api/datasets/1/variables/birthyr/":',
+            '{"name":"Year of birth"},',
+            '"https://app.crunch.io/api/datasets/1/folders/3/":',
+            '{"name":"Nested folder"},',
+            '"https://app.crunch.io/api/datasets/1/variables/textVar/":',
+            '{"name":"FTW! textvar"}}}'
+        )
+    })
+
+    test_that("rename objects inside a folder (validation)", {
+        expect_error(
+            ds %>%
+                cd("Group 1") %>%
+                setNames(c("Year of birth", "Nested folder")),
+            "names must have the same length as the number of children: 3"
+        )
+    })
+
     test_that("mv error handling", {
         expect_error(
             ds %>% cd("Group 1") %>% mv("NOT A VARIABLE", "../Group 2"),
@@ -334,5 +375,10 @@ with_test_authentication({
     test_that("mv preserves order of variables", {
         ds <- mv(ds, c("qsec", "drat", "vs", "wt"), "newdir")
         expect_identical(names(cd(ds, "newdir")), c("qsec", "drat", "vs", "wt"))
+    })
+
+    test_that("setNames works", {
+        ds %>% cd("newdir") %>% setNames(c("queue_sec", "drizzat", "versus", "weight"))
+        expect_identical(names(cd(ds, "newdir")), c("queue_sec", "drizzat", "versus", "weight"))
     })
 })